pyproject.toml
[tool.poetry]
name = "comparative-multimodal-representations"
version = "0.1.0"
description = "Comparative study of multimodal representations in deep learning"


[[tool.poetry.packages]]
include = "comparative"
from = "src"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0"
pytorch-lightning = "^2.0"
hydra-core = "^1.3"
omegaconf = "^2.3"
transformers = "^4.40"
pandas = "^2.0"
numpy = "^1.23"
scikit-learn = "^1.3"
tqdm = "*"
matplotlib = "*"
umap-learn = "*"
networkx = "*"
captum = "*"
grad-cam = "*"
jupyter = "*"
pytest = "*"
wandb = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
